18 reduce code duplication (#20)
* Add header size function, renamed functions

* Change function return types to int64

cleans up warnings

* Refactor map unpacking

* Refactor array unpacking

+++ src/messagepack.f90
@@ -25,7 +25,7 @@ contains
         class(mp_value_type), intent(in) :: obj
         integer, intent(in) :: indentation
         logical, intent(in) :: sameline
-        integer :: i, j
+        integer(kind=int64) :: i, j
 
         if (.not. sameline) then
             do i = 1,indentation
